# Performance Optimization: Replace Linear Property Search with O(1) HashMap Lookup